♻️Refactor : pricing 패키지에 provider 추가(git 누락 추가)
