+++ src/main/java/oridungjeol/duckhang/board/infrastructure/pricing/DelegatePriceProvider.java
@@ -1,9 +1,8 @@
-package oridungjeol.duckhang.board.application.service;
+package oridungjeol.duckhang.board.infrastructure.pricing;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import oridungjeol.duckhang.board.infrastructure.repository.DelegateJpaRepository;
-import oridungjeol.duckhang.board.infrastructure.repository.RentalJpaRepository;
 import oridungjeol.duckhang.payment.domain.PriceProvider;
 
 @Component
